% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateImagePrediction.R
\name{annotateImagePrediction}
\alias{annotateImagePrediction}
\title{Image Annotation}
\usage{
annotateImagePrediction(plotDataOne = plotData[1, ], loc = positiveLoc,
  outputDir = paste0(outputDir, "/", "PositiveImages"))
}
\arguments{
\item{outputDir}{}
}
\description{
Image Annotation
}
\examples{

}
