% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filenameExtractor.R
\name{filenameExtractor}
\alias{filenameExtractor}
\title{Filename Extractor}
\usage{
filenameExtractor(filelist = kmlTiles$Image)
}
\arguments{
\item{filelist}{a vector of filenames that you want to extract lat long values from. Similar to this:
or_goldbeach_20160703--Set2_-123.521_41.94.jpg}
}
\description{
Filename Extractor
}
