% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flightGridKML.R
\name{flightGridKML}
\alias{flightGridKML}
\title{flightGridKML}
\usage{
flightGridKML(rasterDir)
}
\arguments{
\item{rasterDir}{Directory with multiple raster files saved as *.tif.}
}
\value{
flightgrid.kml kml file with raster extents as polygons.
}
\description{
This function takes a directory with raster files and returns a single kml
that contains the extent polygons for all the rasters. Useful to generate a
raster grid to see search region in Google Earth.
}
