% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flightLevelClassifier.R
\name{flightLevelClassifier}
\alias{flightLevelClassifier}
\title{Flight Level Classifier.}
\usage{
flightLevelClassifier(homeDir = getwd(), flightName, modelName, classes,
  positiveClasses = c("TrespassPlants", "TrespassHoles"),
  filterThreshold = 0.2)
}
\arguments{
\item{homeDir}{Base directory.  Should have "Chips" folder.}

\item{flightName}{Name of flight to be classified.}

\item{modelName}{Name of loaded model for labeling.}

\item{classes}{vector of classes for model}
}
\value{

}
\description{
Iterates classification over a flight directory after giving computation
time.
}
\examples{
\dontrun{

}

}
