% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hardNegativeExtractor.R
\name{hardNegativeExtractor}
\alias{hardNegativeExtractor}
\title{Hard negative extractor}
\usage{
hardNegativeExtractor(tileDir = dirList[1], modelName = "M7",
  outputDir)
}
\arguments{
\item{tileDir}{Directory to extract hard negatives from}

\item{modelName}{Model verion name}

\item{outputDir}{Directory to export hard negative to.}
}
\value{
Three random positives from a classified tile copied to new dir
}
\description{
Grabs three random positives from a classified tile.  It can potentially grab
a true positive chip so "negative" is used loosely..
}
\examples{
\dontrun{

}

}
