% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadDIMECModel.R
\name{loadDIMECModel}
\alias{loadDIMECModel}
\title{Title}
\usage{
loadDIMECModel(directory)
}
\arguments{
\item{directory}{}
}
\description{
Title
}
\examples{
\dontrun{
directory <- "F:/Adam Cummings/DimecV1/Model Files/M17"
modM17 <- loadDIMECModel(directory)
}
}
