% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosaicer.R
\name{mosaicer}
\alias{mosaicer}
\title{Raster Mosaicer}
\usage{
mosaicer(homeDir = getwd(), originalFolder = "rgb", flightName,
  mosaicLabel)
}
\arguments{
\item{homeDir}{Base Directory. Mosaic and rgb folders important.}

\item{originalFolder}{Typically "rgb" but could be "original" or googletiles}

\item{flightName}{Name of the flight to mosaic.}

\item{mosaicLabel}{Short memorable}
}
\value{
Exports a mosaiced dataset with 25 small google tiles as one large
 tile
}
\description{
Raster Mosaicer
}
