% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadDIMECModel.R
\name{reloadDIMECModel}
\alias{reloadDIMECModel}
\title{Reload a broken python model pointer in a DIMEC object}
\usage{
reloadDIMECModel(DIMECModel)
}
\arguments{
\item{DIMECModel}{}
}
\description{
Reload a broken python model pointer in a DIMEC object
}
\examples{


}
