% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleClassGrabber.R
\name{singleClassGrabber}
\alias{singleClassGrabber}
\title{Title}
\usage{

  singleClassGrabber(flightDir = "F:/Adam Cummings/GoogleImagery/ca_hayfork_20160529_4096mosaics/Model Output",
  modelName = "M7", class = "PrivateGrow",
  outputDir = "F:/Adam Cummings/GoogleImagery/ca_hayfork_20160529_4096mosaics/Model Summary Output/M7",
  filterRate = 0.8, fileLabel = "Region")
}
\arguments{
\item{flightDir}{Directory of model output to merge for export}

\item{modelName}{Particular model version output to target}

\item{class}{Class name to compile}

\item{outputDir}{Target directory for export}

\item{filterRate}{Threshold value to filter the classifications by.}

\item{fileLabel}{Additional name modifier for exported kml file.}
}
\description{
Title
}
