% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tileLevelClassifier.R
\name{tileLevelClassifier}
\alias{tileLevelClassifier}
\title{Tile Level Classifier.}
\usage{
tileLevelClassifier(tileName, homeDir = getwd(), flightName, modelName,
  exportResults = TRUE, returnPlotData = FALSE, classes = classNames)
}
\arguments{
\item{tileName}{Name of tile to classify}

\item{homeDir}{Base Directory.  Needs "Chips" folder.}

\item{flightName}{Name of flight to classify}

\item{modelName}{Model version to use in classification and labeling.}

\item{exportResults}{Logical.  Option to export results for later analysis}

\item{returnPlotData}{Logical. Option to return plotData from function.}

\item{classes}{Vector of class labels.}
}
\value{
Returns either a data.frame of classifications, or outputs it to a
 plotData.csv, and two kml files, or both. Outputs to "getwd()/Model Output".
  Optionally exports the results as a set of files in directories, including
  positive images, kml's of positives, and the full model results
}
\description{
Important: Requires a model loaded into the environment called "model".
}
