% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topImageExporter.R
\name{topImageExporter}
\alias{topImageExporter}
\title{Top Image Export}
\usage{
topImageExporter(plotData, class = "TrespassPlants", homeDir = getwd(),
  flightName, tileName, modelName, exportDir = "Model Output",
  threshold = 0.9)
}
\arguments{
\item{plotData}{data from which to extract the top images}

\item{class}{Target class}

\item{homeDir}{Base Directory.  Should have Chips folder}

\item{flightName}{Target flight}

\item{tileName}{Target Tile}

\item{modelName}{Name of model file to be used.}

\item{exportDir}{Folder to export the top images to.}

\item{threshold}{threshold for subsetting the results}
}
\value{
Directory.  Creates a directory and populates with images that meet
the prediction threshold.
}
\description{
Top Image Export
}
