% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateImagePrediction.R
\name{annotateImagePrediction}
\alias{annotateImagePrediction}
\title{Image Annotation}
\usage{
annotateImagePrediction(plotDataOne = plotData[1, ],
  imageDir = positiveLoc, outputDir = file.path(outputDir,
  "PositiveImages"))
}
\arguments{
\item{plotDataOne}{A single line to annotate from a plotData file}

\item{imageDir}{Location of images to annotate}

\item{outputDir}{Target directory for new images}
}
\value{
The function exports the provided chips with neat bar graph of
classification.
}
\description{
Image Annotation
}
