% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildBalancedModelData.R
\name{buildBalancedModelData}
\alias{buildBalancedModelData}
\title{Title}
\usage{
buildBalancedModelData(data.folder, trainpercent = 0.8,
  validatepercent = 0.1, testpercent = 0.1)
}
\arguments{
\item{testpercent}{}
}
\description{
Title
}
\examples{


}
