% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildModelData.R
\name{buildModelData}
\alias{buildModelData}
\title{Build Model Data}
\usage{
buildModelData(data.folder = getwd(), trainpercent = 0.85,
  validatepercent = 0.15, testpercent = 0)
}
\arguments{
\item{testpercent}{}
}
\description{
Create train, validation and training data from a pool of original class sorted data
}
\examples{
\dontrun{
buildModelData("F:/Adam Cummings/DimecV1/Data")
}
}
