% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filenameExtractor.R
\name{filenameExtractor}
\alias{filenameExtractor}
\title{Filename Extractor}
\usage{
filenameExtractor(filelist = kmlTiles$Image)
}
\arguments{
\item{filelist}{a vector of filenames that you want to extract lat long values from. Similar to this:
purchase1: or_goldbeach_20160703--Set2_-123.521_41.94.jpg
purchase2: "C01_R01-Hayfork2016_41.1684_-123.5119.jpg"
I flipped lat and long so they can be pasted into google earth easily but now
two different extraction codes are necessary..}
}
\description{
Filename Extractor
}
