% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flightLevelClassifier.R
\name{flightLevelClassifier}
\alias{flightLevelClassifier}
\title{Flight Level Classifier.}
\usage{
flightLevelClassifier(homeDir = getwd(), modelObj, flightName,
  filterThreshold = 0.2, assessFlightSize = TRUE)
}
\arguments{
\item{homeDir}{Base directory.  Should have "Chips" folder.}

\item{modelObj}{A DIMEC model object.}

\item{flightName}{Name of flight to be classified.}

\item{filterThreshold}{Adjustable threshold for classifying a positive hit.}

\item{assessFlightSize}{Spend a couple minutes getting a precise time estimate?}
}
\value{

}
\description{
Iterates classification over a flight directory after giving computation
time.
}
\examples{
\dontrun{

}

}
