% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flightLevelClassifier.R
\name{flightLevelClassifier}
\alias{flightLevelClassifier}
\title{Flight Level Classifier.}
\usage{
flightLevelClassifier(flightDir = dirList[1], modelName = modelLabel,
  exportResults = TRUE, returnPlotData = FALSE, classes = classNames)
}
\arguments{
\item{flightDir}{Directory of the chips to classify}

\item{modelName}{Model version to import for classification}

\item{exportResults}{Logical of whether you want the export the classificaiton results to the "Model Output" directory}

\item{returnPlotData}{Logical of whether you want to return the plotData classification data.}

\item{classes}{Vector of class names.}
}
\value{
Returns either a data.frame of classifications, or outputs it to a plotData.csv, and two kml files, or both. Outputs to "getwd()/Model Output".

DirectoryofResults Optionally exports the results as a set of files in directories, including positive images, kml's of positives, and the full model results
}
\description{
Flight Level Classifier.
}
\examples{
\dontrun{

}

}
