% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flightLevelClassifier.R
\name{flightLevelClassifier}
\alias{flightLevelClassifier}
\title{Title}
\usage{
flightLevelClassifier(flightDir = dirList[1], modelName = modelName,
  exportResults = TRUE, returnPlotData = FALSE)
}
\arguments{
\item{returnPlotData}{}
}
\description{
Title
}
\examples{


}
