% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flightMetricTally.R
\name{flightMetricTally}
\alias{flightMetricTally}
\title{Flight Metric Tally}
\usage{
flightMetricTally(homeDir = getwd(), modelObj, flightName,
  fileName = NULL, summariseKMLs = FALSE)
}
\arguments{
\item{homeDir}{Base Directory. Should have Model Output Directory.}

\item{modelObj}{A DIMEC Model object.}

\item{flightName}{Name of flight to summarize}

\item{fileName}{Output file name.  Be sure to add .csv to end.}

\item{summariseKMLs}{Logical. If true, calls kmlCompiler to produce summary
kmls}

\item{positiveClasses}{Vector of class labels that are considered positive
hits}
}
\value{
fileName.csv This file summarises the predictions at each tile for an
entire flight. The file contains several metrics of determining a "positive"
but they all seem about equivalent just with lower or higher sensitivity.

kmls optionally the function can export flight-level merged kmls
by calling kmlCompiler
}
\description{
Flight Metric Tally
}
\examples{
\dontrun{
flightMetricTally(modelName="M17",
   flightName = "ca_dosrios_20170813")
}

}
