% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flightMetricTally.R
\name{flightMetricTally}
\alias{flightMetricTally}
\title{Flight Metric Tally}
\usage{
flightMetricTally(flightDir = "Model Output", modelName = "M14",
  outputDir = "Model Summary Output", fileName = "MetricTally.csv")
}
\arguments{
\item{flightDir}{Directory of multiple tile outputs}

\item{modelName}{DIMEC Model version to extract}

\item{outputDir}{Target location for output}

\item{fileName}{Output file name.  Be sure to add .csv to end.}
}
\value{
outputs a csv file that contains tallys of positive detections at different thresholds.
}
\description{
Flight Metric Tally
}
\examples{
\dontrun{
flightMetricTally(modelName="M14")
}
}
