% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmlMaker.R
\name{kmlMaker}
\alias{kmlMaker}
\title{KML exporter}
\usage{
kmlMaker(kmlExportData = plotData, filename = "exportedkml",
  exportDir = getwd(), layerName = "ModelOutput")
}
\arguments{
\item{kmlExportData}{Requires columns named Latitude and Longitude in the epsg:4326 projection.  all other columns added to popup bubble of kml points.}

\item{layerName}{}
}
\description{
This function exports a kml into the designated directory
}
