% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retile.R
\name{retile}
\alias{retile}
\title{Title}
\usage{
retile(imageName, flightName = "ca_hayfork_20160529_rgb", red = 1,
  green = 2, blue = 3, dim = 299, overlap = 60,
  outputFolder = "Chips", outputJPG = TRUE, returnData = FALSE)
}
\arguments{
\item{imageName}{Name of tiff file to chip.}

\item{red}{Dimension number for red band}

\item{green}{Dimension number for green band}

\item{blue}{Dimension number for blue band}

\item{dim}{Size of output jpg chip}

\item{overlap}{number of pixels to overlap the chips}

\item{outputFolder}{location that new chips are copied}

\item{imageDir}{Folder for current operations}

\item{inputFolder}{Folder location of tiff file}
}
\description{
Title
}
\examples{
\dontrun{
###List of Tiffs to Crop:-----------------------------------------------
files <- list.files("G:/GoogleImagery/Purchase2/Mosaics",recursive=TRUE,pattern="*.tif",full.names = TRUE)
files <- files[grep(pattern = "gold",x = files)]

###Crop one tiff:--------------------------------------------
retile(imageName = files[1],dim=299,overlap=60,outputFolder="Chips")

###Crop a list of tiffs:-------------------------------------------
sapply(files,FUN=retile,outputFolder="Chips")

}
}
