% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retile.R
\name{retile}
\alias{retile}
\title{Title}
\usage{
retile(imageDir = getwd(), inputFolder = "Mosiacs", imageName,
  red = 1, green = 2, blue = 3, dim = 299, overlap = 60,
  outputFolder = "Tiles")
}
\arguments{
\item{imageDir}{Folder for current operations}

\item{inputFolder}{Folder location of tiff file}

\item{imageName}{Name of tiff file to chip.}

\item{red}{Dimension number for red band}

\item{green}{Dimension number for green band}

\item{blue}{Dimension number for blue band}

\item{dim}{Size of output jpg chip}

\item{overlap}{number of pixels to overlap the chips}

\item{outputFolder}{location that new chips are copied}
}
\description{
Title
}
\examples{
\dontrun{
###List of Tiffs to Crop:-----------------------------------------------
files <- list.files("F:/Adam Cummings/GoogleImagery/ShastaT/Mosaics",recursive=FALSE)
files <- files[grep(pattern = "gold",x = files)]

###Crop one tiff:--------------------------------------------
retile("F:/Adam Cummings/GoogleImagery/ShastaT",
       imageName = files[2],
       dim=299,overlap=60,outputFolder="Chips",inputFolder="Mosaics")

###Crop a list of tiffs:-------------------------------------------
sapply(files,FUN=retile,
       imageDir = "G:/ShastaT Search/ca_arcata_20160527",
       dim=299,overlap=60,outputFolder="Chips",inputFolder="Mosaics")

}
}
