% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleClassGrabber.R
\name{singleClassGrabber}
\alias{singleClassGrabber}
\title{Title}
\usage{

  singleClassGrabber(flightDir = "F:/Adam Cummings/GoogleImagery/ca_hayfork_20160529_4096mosaics/Model Output",
  modelName = "M7", class = "PrivateGrow",
  outputDir = "F:/Adam Cummings/GoogleImagery/ca_hayfork_20160529_4096mosaics/Model Summary Output/M7",
  filterRate = 0.8, fileLabel = "Region")
}
\arguments{
\item{fileLabel}{}
}
\description{
Title
}
