% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleClassGrabber.R
\name{singleClassGrabber}
\alias{singleClassGrabber}
\title{Single Class Grabber}
\usage{
singleClassGrabber(homeDir = getwd(), flightName, modelName = "M7",
  class = "PrivateGrow", filterRate = 0, fileLabel = "Region")
}
\arguments{
\item{homeDir}{Base Directory. Should have Model Output and Model Output
Summary}

\item{flightName}{flight name model output to merge for export}

\item{modelName}{Particular model version output to target}

\item{class}{Class name to compile}

\item{filterRate}{Predictions lower than this value are dropped. Default is
no threshold.}

\item{fileLabel}{Additional name modifier for exported kml file.}
}
\value{
saves a kml file in the "Model Output Summary" for the flight..
}
\description{
This function searches across an entire flight and pulls classifications that
match a given class.  Use this to pull all greenhouses, for example.
}
