% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topImageExporter.R
\name{topImageExporter}
\alias{topImageExporter}
\title{Top Image Export}
\usage{
topImageExporter(plotData, class = "TrespassPlants",
  unclassifieddir = searchdir, exportdir = exportdir,
  threshold = 0.9)
}
\arguments{
\item{plotData}{data from which to extract the top images}

\item{class}{Target class}

\item{unclassifieddir}{location of chips}

\item{exportdir}{target directory for new files}

\item{threshold}{threshold for exporting the top images}
}
\description{
Top Image Export
}
\examples{

}
